% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otd.r
\name{otd}
\alias{otd}
\title{Metadata}
\usage{
otd(dsn, layer = NULL, stringsAsFactors = FALSE, integer64 = "no.loss",
  verbose = TRUE)
}
\arguments{
\item{dsn}{data source name, i.e. a file path}

\item{layer}{named layer, numbered layer, or left otu}

\item{stringsAsFactors}{never}

\item{integer64}{policy on these}

\item{verbose}{verbosity TRUE or FALSE}
}
\value{
a data frame, no factors is the default
}
\description{
Just the metadata, for your interpretatation.
}
\details{
"Metadata" here means "feature level" data, the rows in the spatial table.
}
\examples{
inlandwaters <- otd(system.file("extdata", "datasource.gpkg", package = "gladr"))
##mixed_poly_lines <- otd(system.file("extdata", "Drawing.mif", package = "gladr"))
}
