% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otg.R
\name{otg}
\alias{otg}
\title{Geometry}
\usage{
otg(x, y = NULL, verbose = TRUE)
}
\arguments{
\item{x}{data source name}

\item{y}{layer name, or number, defaults to first layer '1'}

\item{verbose}{report on action if \code{TRUE} and \code{y} for layer not specified, the available layers will be reported}
}
\value{
raw list as per 'R_OGR_CAPI_FEATURES'
}
\description{
Just the geometry, for your interpretation.
}
\details{
This function returns the raw output of the internal source function
'R_OGR_CAPI_features' in 'rgdal'.
It is possible to have the same layer name in a folder of files, i.e
for MapInfo or shapefile, so don't do that.
}
\examples{
dsn <- system.file("extdata", "mapinfo", package = "otg")
g1 <- otg(dsn)

#dsn <- system.file("extdata", "inlandwaters.gpkg", package = "ggpolypath")

mt <- otg:::mtg(otg::otg(system.file("extdata", "inlandwaters.gpkg", package = "ggpolypath")))
plot(spbabel::sp(mt))
 plot(spbabel::sp(mt)[3, ], asp = "", col = "grey", rule = 'evenodd')
}

